fixed QPM output QPM is temporarily called QPH (until more search requests are present?)
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@3313 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/yacy/yacySeed.java
@@ -433,11 +433,11 @@ public class yacySeed {
         }
     }
 
-    public int getQPM() {
+    public double getQPM() {
         try {
-            return Integer.parseInt(get(yacySeed.RSPEED, yacySeed.ZERO));
+            return Double.parseDouble(get(yacySeed.RSPEED, yacySeed.ZERO));
         } catch (NumberFormatException e) {
-            return 0;
+            return 0d;
         }
     }
